requirements.txt
asgiref==3.3.1
attrs==20.3.0
cffi==1.14.5
cryptography==3.4.7
decorator==4.4.2
Django==3.1.6
django-appconf==1.0.4
gunicorn==20.0.4
importlab==0.6.1
networkx==2.5
ninja==1.10.0.post2
Pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
pyperclip==1.8.2
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
sqlparse==0.4.1
typed-ast==1.4.2
whitenoise==5.2.0
